requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
asyncio>=3.4.3
pytz>=2021.3
aiosqlite>=0.19.0
